fix warnings

+++ exp/apps/cuthill/cuthill-distr.cpp
@@ -651,7 +651,7 @@ struct banddiff {
 			GNode dst = graph.getEdgeDst(ii);
 			SNode& ddata = graph.getData(dst, Galois::MethodFlag::UNPROTECTED);
 
-			long int diff = abs(sdata.id - ddata.id);
+			long int diff = abs(static_cast<long int>(sdata.id) - static_cast<long int>(ddata.id));
 			//long int diff = (long int) sdata.id - (long int) ddata.id;
 			maxdiff = diff > maxdiff ? diff : maxdiff;
 		}
